refactored for a2c
