+++ src/ca/bcit/comp2526/a2c/RandomGenerator.java
@@ -1,4 +1,4 @@
-package ca.bcit.comp2526.a2b;
+package ca.bcit.comp2526.a2c;
 
 /**
  * Pseudo-random number generator.
